+++ pyhash.h
@@ -24,11 +24,11 @@ typedef enum {
 void    inithash (void);
 void    closehash (void);
 void	HashStats(int level, char *trailer);
-boolean inhash(hashwhat what, smallint val);
-void addtohash(hashwhat what, smallint val);
-boolean ser_dsrsol(couleur camp, smallint n, boolean restartenabled);
-boolean shsol(couleur camp, smallint n, boolean restartenabled);
-boolean mataide(couleur camp, smallint n, boolean restartenabled);
+boolean inhash(hashwhat what, int val);
+void addtohash(hashwhat what, int val);
+boolean ser_dsrsol(couleur camp, int n, boolean restartenabled);
+boolean shsol(couleur camp, int n, boolean restartenabled);
+boolean mataide(couleur camp, int n, boolean restartenabled);
 boolean last_dsr_move(couleur camp);
 
 #endif
