+++ pyproof.h
@@ -11,8 +11,8 @@
 
 BCMemValue *ProofEncode(void);
 void    ProofInitilise(void);
-boolean ProofSol(couleur camp, smallint n, boolean restartenabled);
-boolean SeriesProofSol(smallint n, boolean restartenabled);
+boolean ProofSol(couleur camp, int n, boolean restartenabled);
+boolean SeriesProofSol(int n, boolean restartenabled);
 void    ProofInitialiseKingMoves(square ProofRB, square ProofRN);
 boolean ProofVerifie(void);
 void    ProofInitialise(void);
